# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BizManager Purchase & Sales Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: List products
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: lowStock
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /products/{id}/adjust:
    post:
      operationId: adjustStock
      tags: [products]
      summary: Adjust product stock
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockAdjustRequest"
      responses:
        "200":
          description: Stock adjusted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /purchases:
    get:
      operationId: getPurchases
      tags: [purchases]
      summary: List purchases
      security:
        - bearerAuth: []
      parameters:
        - name: supplierId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of purchases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Purchase"
    post:
      operationId: createPurchase
      tags: [purchases]
      summary: Create purchase
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePurchaseRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
  /purchases/{id}:
    get:
      operationId: getPurchase
      tags: [purchases]
      summary: Get purchase
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
...
  /returns:
    get:
      operationId: getReturns
      tags: [returns]
      summary: List returns
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of returns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Return"
    post:
      operationId: createReturn
      tags: [returns]
      summary: Create return
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReturnRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Return"
...
  /reports/profit-loss:
    get:
      operationId: getProfitLossReport
      tags: [reports]
      summary: Get profit and loss report
      security:
        - bearerAuth: []
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: groupBy
          in: query
          schema:
            type: string
            enum: [overall, product, rack]
      responses:
        "200":
          description: P&L report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfitLossReport"
  /reports/stock:
    get:
      operationId: getStockReport
      tags: [reports]
      summary: Get stock report
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Stock report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockReportItem"
  /reports/top-products:
    get:
      operationId: getTopProducts
      tags: [reports]
      summary: Get top selling products
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Top products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopProduct"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        purchasePrice:
          type: number
        sellingPrice:
          type: number
        gstPercent:
          type: number
        rack:
          type: string
        stockQty:
          type: number
        lowStockThreshold:
          type: number
        unit:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - purchasePrice
        - sellingPrice
        - gstPercent
        - stockQty
        - createdAt
    CreateProductRequest:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        purchasePrice:
          type: number
        sellingPrice:
          type: number
        gstPercent:
          type: number
        rack:
          type: string
        stockQty:
          type: number
        lowStockThreshold:
          type: number
        unit:
          type: string
      required:
        - name
        - purchasePrice
        - sellingPrice
        - gstPercent
    StockAdjustRequest:
      type: object
      properties:
        adjustment:
          type: number
        reason:
          type: string
      required:
        - adjustment
    PurchaseItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: number
        purchasePrice:
          type: number
        gstPercent:
          type: number
        gstAmount:
          type: number
        totalAmount:
          type: number
      required:
        - productId
        - productName
        - quantity
        - purchasePrice
        - gstPercent
        - gstAmount
        - totalAmount
    Purchase:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        supplierId:
          type: integer
        supplierName:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseItem"
        subtotal:
          type: number
        totalGst:
          type: number
        grandTotal:
          type: number
        paymentMode:
          type: string
        notes:
          type: string
        date:
          type: string
        createdAt:
          type: string
      required:
        - id
        - invoiceNumber
        - supplierId
        - supplierName
        - items
        - subtotal
        - totalGst
        - grandTotal
        - paymentMode
        - date
        - createdAt
    CreatePurchaseRequest:
      type: object
      properties:
        supplierId:
          type: integer
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: integer
              quantity:
                type: number
              purchasePrice:
                type: number
              gstPercent:
                type: number
            required:
              - productId
              - quantity
              - purchasePrice
              - gstPercent
        paymentMode:
          type: string
          enum: [cash, credit, upi, bank]
        notes:
          type: string
        date:
          type: string
      required:
        - supplierId
        - items
        - paymentMode
...
    Return:
      type: object
      properties:
        id:
          type: integer
        returnType:
          type: string
        referenceId:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: number
        amount:
          type: number
        refundMode:
          type: string
        reason:
          type: string
        date:
          type: string
        createdAt:
          type: string
      required:
        - id
        - returnType
        - referenceId
        - productId
        - productName
        - quantity
        - amount
        - date
        - createdAt
    CreateReturnRequest:
      type: object
      properties:
        returnType:
          type: string
          enum: [purchase_return, sale_return]
        referenceId:
          type: integer
        productId:
          type: integer
        quantity:
          type: number
        amount:
          type: number
        refundMode:
          type: string
          enum: [cash, upi, credit, none]
        reason:
          type: string
        date:
          type: string
      required:
        - returnType
        - referenceId
        - productId
        - quantity
        - amount
        - refundMode
...
    ProfitLossReport:
      type: object
      properties:
        totalRevenue:
          type: number
        totalCost:
          type: number
        grossProfit:
          type: number
        profitMargin:
          type: number
        items:
          type: array
          items:
            type: object
            properties:
              label:
                type: string
              revenue:
                type: number
              cost:
                type: number
              profit:
                type: number
              margin:
                type: number
            required:
              - label
              - revenue
              - cost
              - profit
              - margin
      required:
        - totalRevenue
        - totalCost
        - grossProfit
        - profitMargin
        - items
    StockReportItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        rack:
          type: string
        stockQty:
          type: number
        lowStockThreshold:
          type: number
        purchasePrice:
          type: number
        sellingPrice:
          type: number
        stockValue:
          type: number
        isLowStock:
          type: boolean
      required:
        - id
        - name
        - stockQty
        - purchasePrice
        - sellingPrice
        - stockValue
        - isLowStock
    TopProduct:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        totalSold:
          type: number
        totalRevenue:
          type: number
        totalProfit:
          type: number
      required:
        - id
        - name
        - totalSold
        - totalRevenue
        - totalProfit